+++ Optimizer/ErrorLogger.cs
@@ -13,16 +13,7 @@ namespace Optimizer
             {
                 if (!File.Exists(ErrorLogFile) || (File.Exists(ErrorLogFile) && File.ReadAllText(ErrorLogFile).Trim() == string.Empty))
                 {
-                    string bitness = Environment.Is64BitOperatingSystem ? "64-bit" : "32-bit";
-
-                    if (Utilities.CurrentWindowsVersion == WindowsVersion.Windows10)
-                    {
-                        File.AppendAllText(ErrorLogFile, string.Format("{0} - {1} ({2})", Utilities.GetOS(), Utilities.GetWindows10Build(), bitness));
-                    }
-                    else
-                    {
-                        File.AppendAllText(ErrorLogFile, string.Format("{0} - ({1})", Utilities.GetOS(), bitness));
-                    }
+                    File.AppendAllText(ErrorLogFile, Utilities.GetWindowsDetails());
                     File.AppendAllText(ErrorLogFile, Environment.NewLine);
                     File.AppendAllText(ErrorLogFile, string.Format("Optimizer {0} - .NET Framework {1} - Experimental build: {2}", Program.GetCurrentVersionTostring(), Utilities.GetNETFramework(), Program.EXPERIMENTAL_BUILD));
                     File.AppendAllText(ErrorLogFile, Environment.NewLine);
@@ -43,6 +34,13 @@ namespace Optimizer
                 File.AppendAllText(ErrorLogFile, Environment.NewLine);
             }
             catch { }
+            finally
+            {
+                if (!Options.CurrentOptions.DisableOptimizerTelemetry)
+                {
+                    TelemetryHelper.GenerateTelemetryData(functionName, errorMessage, errorStackTrace);
+                }
+            }
         }
     }
 }